Task D

+++ MP5/src/RandomForestMP.java
@@ -21,9 +21,9 @@ public final class RandomForestMP {
         public LabeledPoint call(String line) throws Exception {
             String[] tok = SPACE.split(line);
             int label = Integer.parseInt(tok[tok.length-1]);
-            int[] point = new int[tok.length-1];
+            double[] point = new double[tok.length-1];
             for (int i = 0; i < tok.length - 1; ++i) {
-                point[i] = Integer.parseInt(tok[i]);
+                point[i] = Double.parseDouble(tok[i]);
             }
             return new LabeledPoint(label, Vectors.dense(point));
         }
